+++ src/components/bottom-sheets/index.ts
@@ -1,4 +1,5 @@
 export { default as ActionBottomSheet } from './ActionBottomSheet';
 export { default as LocationQueryBottomSheet } from './LocationQueryBottomSheet';
+export { default as SelectFromLibraryBottomSheet } from './SelectFromLibraryBottomSheet';
 
 export type { ActionBottomSheetItem } from './ActionBottomSheet';