+++ src/constants/values.ts
@@ -6,9 +6,11 @@ const nodePackage = require('../../package.json');
 
 export const DEFAULT_ACTIVE_OPACITY = 0.75;
 export const DEFAULT_TILE_SPACING = spacing.sm * 1.1;
-export const DEFAULT_MIN_BOTTOM_TAB_BAR_HEIGHT = 55;
-
+export const DEFAULT_MIN_BOTTOM_TAB_BAR_HEIGHT = 54;
 export const DEFAULT_ICON_LIKE_ANIMATION: Animation = 'rubberBand';
+
+export const MAX_VID_DURATION_SECONDS = 60;
+export const MAX_VID_DURATION_MILLISECONDS = MAX_VID_DURATION_SECONDS * 1000;
 
 export const APP_VERSION: string =
   nodePackage.version || `${getVersion()}-native` || '<unknown-version>';
@@ -23,9 +25,6 @@ function createVersionString(
 ): string {
   const [major, minor, patch, build] = version;
   return String(
-    major * 10 ** 6 +
-      (minor % 100) * 10 ** 4 +
-      (patch % 100) * 10 ** 2 +
-      (build % 100),
+    major * 1e6 + (minor % 100) * 1e4 + (patch % 100) * 1e2 + (build % 100),
   );
 }