+++ src/features/create/components/VideoPreviewPicker.tsx
@@ -18,12 +18,14 @@ import { useFocusEffect } from '@react-navigation/core';
 
 import * as utilities from 'src/utilities';
 import { ActionBottomSheet, ActionBottomSheetItem } from 'src/components';
-import { color, layout } from 'src/constants';
+
+import * as constants from 'src/constants';
+import {
+  MAX_VID_DURATION_MILLISECONDS,
+  MAX_VID_DURATION_SECONDS,
+} from 'src/constants/values';
 
 import PreviewPicker, { PreviewPickerProps } from './PreviewPicker';
-
-const MAXIMUM_DURATION_SECONDS = 60;
-const MAXIMUM_DURATION_MILLISECONDS = MAXIMUM_DURATION_SECONDS * 1000;
 
 // const command = `-i ${video.path} -loop -1 -vf "scale=320:-1" ${outputPath}`;
 
@@ -92,13 +94,13 @@ export default function ViewPreviewPicker(props: VideoPreviewPickerProps) {
             selectedVideos.some(
               video =>
                 video.duration &&
-                video.duration > MAXIMUM_DURATION_MILLISECONDS,
+                video.duration > MAX_VID_DURATION_MILLISECONDS,
             )
           ) {
             Alert.alert(
               'Video Will Be Trimmed',
-              'One or more of your videos are longer than the maximum time ' +
-                `limit of ${MAXIMUM_DURATION_SECONDS} seconds. It will be ` +
+              'One or more of your videos are longer than the maximum video ' +
+                `duration of ${MAX_VID_DURATION_SECONDS} seconds. It will be ` +
                 'trimmed when you upload it.',
             );
           }
@@ -113,12 +115,12 @@ export default function ViewPreviewPicker(props: VideoPreviewPickerProps) {
 
           if (
             selectedVideo.duration &&
-            selectedVideo.duration > MAXIMUM_DURATION_MILLISECONDS
+            selectedVideo.duration > MAX_VID_DURATION_MILLISECONDS
           ) {
             Alert.alert(
               'Video Will Be Trimmed',
-              'This video is longer than the maximum time limit of ' +
-                `${MAXIMUM_DURATION_SECONDS} seconds. It will be trimmed ` +
+              'This video is longer than the maximum video duration of ' +
+                `${MAX_VID_DURATION_SECONDS} seconds. It will be trimmed ` +
                 'when you upload it.',
             );
           }
@@ -221,15 +223,15 @@ const VideoPreviewPickerItem = React.forwardRef<
         style={{
           width: itemWidth,
           aspectRatio: 1,
-          borderRadius: layout.radius.md,
+          borderRadius: constants.layout.radius.md,
           overflow: 'hidden',
-          backgroundColor: color.placeholder,
+          backgroundColor: constants.color.placeholder,
         }}
       />
       {isLoading && (
         <ActivityIndicator
           size="large"
-          color={color.gray500}
+          color={constants.color.gray500}
           style={{ position: 'absolute', transform: [{ scale: 1.5 }] }}
         />
       )}